SOV-2613: General layout and folder structure (#433)
* General layout and folder structure

* Create healthy-frogs-listen.md

* Fix components folder space

* chore: update changeset message

---------
