+++ .changeset/healthy-frogs-listen.md
@@ -1,0 +1,5 @@
+---
+"frontend": patch
+---
+
+SOV-2613: Reward page - general layout and folder structure
