+++ apps/frontend/src/app/5_pages/RewardsPage/RewardsPage.module.css
@@ -1,0 +1,11 @@
+.desktop {
+  @apply hidden lg:block;
+}
+
+.mobile {
+  @apply block lg:hidden;
+}
+
+.mobileSelect {
+  @apply flex mt-6 justify-center lg:hidden;
+}
