+++ apps/frontend/src/app/5_pages/RewardsPage/components/Staking.tsx
@@ -1,0 +1,7 @@
+import React, { FC } from 'react';
+
+export const Staking: FC = () => (
+  <>
+    <div>TBD - Staking rewards table</div>
+  </>
+);
