+++ apps/frontend/src/app/5_pages/RewardsPage/components/Vesting.tsx
@@ -1,0 +1,7 @@
+import React, { FC } from 'react';
+
+export const Vesting: FC = () => (
+  <>
+    <div>TBD - Vesting rewards table</div>
+  </>
+);
