+++ apps/frontend/src/app/5_pages/RewardsPage/hooks/useHandleRewards.ts
@@ -33,16 +33,16 @@ export const useHandleRewards = (
   const title = useMemo(
     () =>
       isWithdrawTransaction
-        ? t(translations.rewardPage.tx.withdrawGains)
-        : t(translations.rewardPage.tx.transferToLOC),
+        ? t(translations.rewardPage.stabilityPool.tx.withdrawGains)
+        : t(translations.rewardPage.stabilityPool.tx.transferToLOC),
     [isWithdrawTransaction],
   );
 
   const transactionTitle = useMemo(
     () =>
       isWithdrawTransaction
-        ? t(translations.rewardPage.tx.withdraw)
-        : t(translations.rewardPage.tx.transfer),
+        ? t(translations.rewardPage.stabilityPool.tx.withdraw)
+        : t(translations.rewardPage.stabilityPool.tx.transfer),
     [isWithdrawTransaction],
   );
